Cleanup: embeddings are Float, cached as packed Data
The on-device vectors were [Double] — twice the model's native precision
and twice the storage. Now:

- EmbeddingProvider.vector returns [Float]; Embedding cosine/normalize are
  Float (vDSP). MiniLM reads the model's Float32 output directly instead of
  widening every element to Double.
- JournalistProfile.embedding is a computed [Float] over a private
  embeddingBlob: Data (raw little-endian Float32, 1536 bytes) — SwiftData
  won't persist a [Float] array, and packed Data is tighter than [Double]
  would have been anyway. Device-local cache, re-warmed on a dimension
  mismatch, so byte-order portability is a non-issue.

Eval output unchanged. Schema change (embedding column) — parkStore handles
it, pre-1.0. 76 unit + UI smoke green.

+++ BeatMatch/Sources/Services/MatchRunner.swift
@@ -26,8 +26,8 @@ enum MatchRunner {
             .map { ($0.persistentModelID, $0.embeddingText) }
         guard !pending.isEmpty else { return 0 }
 
-        let warmed: [PersistentIdentifier: [Double]] = await Task.detached(priority: priority) {
-            var out: [PersistentIdentifier: [Double]] = [:]
+        let warmed: [PersistentIdentifier: [Float]] = await Task.detached(priority: priority) {
+            var out: [PersistentIdentifier: [Float]] = [:]
             for item in pending {
                 if let v = embeddings.vector(for: item.text),
                    v.count == MiniLMEmbeddingProvider.dimension {
@@ -71,7 +71,7 @@ enum MatchRunner {
         let pool = JournalistDirectory.matchable(context)
 
         // The story's own vector — also off the main actor.
-        let storyVector: [Double]? = await Task.detached(priority: .userInitiated) {
+        let storyVector: [Float]? = await Task.detached(priority: .userInitiated) {
             embeddings?.vector(for: Embedding.storyText(analysis, rawText: rawText))
         }.value
